pnews-poller: first refactor towards CLI app
This is a first commit as part of an effort to move away from
the infinite loop + timeout approach that is currently used in
pnews-poller.
Instead, we're moving towards a simple CLI app for pnews-poller that is
periodically invoked by cron.

More background on why this is better:
http://stackoverflow.com/questions/19393915/
what-are-pros-cons-of-cron-job-over-an-infinite-loop-running

In particular, this commit:
- Adds commander.js for easy CLI app creation
- Allows pnews-poller to be invoked with either a --source argument
  specifying
  reddit or hackernews.
- We now properly close the MongoDB connection when we're done saving to
  the DB
- More informative logging in the hackernews poller
- Minor README update