pnews-poller is now periodically ran using cron
Before pnews-poller was run as a service which periodically polled news
sources. This commit removes the service part and introduces a cron job
that runs once a minute instead. The benefit of this is that the
periodic cron job uses less system resources between runs and is more
resilient to errors when a single pnews-poller run errors out.